#10 refactor : card-detail
service 상속을 통한 버전 분리
